Backend/requirements.txt
Flask
flask-cors
polars
pandas
openpyxl 
mysql-connector-python
python-dotenv
xlsxwriter 